(Data Research Project Commit 2)
What I changed
(I added tables that presented the datasets that I have chosen in average and also its normal values. I have also finished the code for the subquestion and have made some progress on the third one.)

Why I changed it
(The reason why I decided to add the averages and normal values in the table format is so that I could made calculations by referring to the table, as I believe that by doing this it makes it easier. I also finished my code because I wanted to get closer to answering my over arching question.)